requirements.txt
# Django Core
Django>=5.0,<5.1
python-decouple>=3.8

# Database
psycopg2-binary>=2.9.9

# Authentication & Permissions
# Using Django's built-in auth system

# Audit Trail & Soft Delete
django-safedelete>=1.3.0
django-simple-history>=3.4.0

# Forms & UI
django-crispy-forms>=2.1
crispy-tailwind>=1.0.3
django-filter>=23.5

# File Handling
Pillow>=10.1.0

# Utilities
django-extensions>=3.2.3

# Development
django-debug-toolbar>=4.2.0
ipython>=8.18.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
